+++ R/radhy.R
@@ -40,7 +40,13 @@ radhy <- function(ADSL, # nolint
                     "TBILI > 2 times ULN and two consecutive elevations of ALT in relation to Baseline",
                     "TBILI <= 2 times ULN and two consecutive elevations of ALT in relation to Baseline",
                     "TBILI > 2 times ULN and two consecutive elevations of AST in relation to Baseline",
-                    "TBILI <= 2 times ULN and two consecutive elevations of AST in relation to Baseline"
+                    "TBILI <= 2 times ULN and two consecutive elevations of AST in relation to Baseline",
+                    "ALT > 3 times ULN by Period",
+                    "AST > 3 times ULN by Period",
+                    "ALT or AST > 3 times ULN by Period",
+                    "ALT > 3 times Baseline by Period",
+                    "AST > 3 times Baseline by Period",
+                    "ALT or AST > 3 times Baseline by Period"
                   ),
                   paramcd = c(
                     "BLAL",
@@ -58,7 +64,13 @@ radhy <- function(ADSL, # nolint
                     "BG2AL2CB",
                     "BL2AL2CB",
                     "BG2AS2CB",
-                    "BL2AS2CB"
+                    "BL2AS2CB",
+                    "ALTPULN",
+                    "ASTPULN",
+                    "ALTASTPU",
+                    "ALTPBASE",
+                    "ASTPBASE",
+                    "ALTASTPB"
                   ),
                   seed = NULL,
                   cached = FALSE) {
@@ -85,46 +97,58 @@ radhy <- function(ADSL, # nolint
     USUBJID = ADSL$USUBJID,
     PARAM = as.factor(param_init_list$relvar1),
     AVISIT = as.factor(c("BASELINE", "POST-BASELINE")),
+    APERIODC = as.factor(c("PERIOD 1", "PERIOD 2")),
     stringsAsFactors = FALSE)
 
-  # define parameters that will be assigned values "Y" or "N"
-  paramcd_yn <- c("BL2AL2CU", "BG2AS2CU", "BL2AS2CU", "BG2AL2CU", "BG2AL2CB", "BL2AL2CB", "BG2AS2CB", "BL2AS2CB") # nolint
+  # remove records that are not needed and were created as a side product of expand.grid above
+  ADHY <- dplyr::filter(ADHY, !(.data$AVISIT == "BASELINE" & .data$APERIODC == "PERIOD 2")) # nolint
 
-  # Add other variables to ADHY
+  # define TBILI ALT/AST params, period dependent parameters and the parameters that will be assigned values "Y" or "N"
+  paramcd_tbilialtast <- c("BLAL", "BGAS", "BGAL", "BLAS", "BA2AL", "BA2AS", "BA5AL", "BA5AS") # nolint
+  paramcd_by_period <- c("ALTPULN", "ASTPULN", "ALTASTPU", "ALTPBASE", "ASTPBASE", "ALTASTPB")
+  paramcd_yn <- c(
+    "BL2AL2CU", "BG2AS2CU", "BL2AS2CU", "BG2AL2CU", "BG2AL2CB", "BL2AL2CB", "BG2AS2CB", "BL2AS2CB",
+    paramcd_by_period
+  ) # nolint
+
+  # add other variables to ADHY
   ADHY <- ADHY %>% # nolint
     dplyr::mutate(
       PARAMCD = factor(rel_var(
         df = as.data.frame(ADHY),
         var_values = param_init_list$relvar2,
         related_var = "PARAM")),
+      AVALC = dplyr::case_when(
+        .data$PARAMCD %in% paramcd_tbilialtast ~ sample(
+          x = c(">3-5ULN", ">5-10ULN", ">10-20ULN", ">20ULN", "Criteria not met"), size = dplyr::n(), replace = TRUE),
+        .data$PARAMCD %in% paramcd_yn ~ sample(
+          x = c("Y", "N"), prob = c(0.1, 0.9), size = dplyr::n(), replace = TRUE)
+        ),
+      AVAL = dplyr::case_when(
+        .data$AVALC == ">3-5ULN" ~ 1,
+        .data$AVALC == ">5-10ULN" ~ 2,
+        .data$AVALC == ">10-20ULN" ~ 3,
+        .data$AVALC == ">20ULN" ~ 4,
+        .data$AVALC == "Y" ~ 1,
+        .data$AVALC == "N" ~ 0,
+        .data$AVALC == "Criteria not met" ~ 0
+      ),
       AVISITN = dplyr::case_when(
-        .data$AVISIT == "BASELINE" ~ 0,
-        .data$AVISIT == "POST-BASELINE" ~ 9995,
-        TRUE ~ NA_real_),
+        .data$AVISIT == "BASELINE" ~ 0L,
+        .data$AVISIT == "POST-BASELINE" ~ 9995L,
+        TRUE ~ NA_integer_),
+      APERIOD = dplyr::case_when(
+        .data$APERIODC == "PERIOD 1" ~ 1L,
+        .data$APERIODC == "PERIOD 2" ~ 2L,
+        TRUE ~ NA_integer_),
       ABLFL = dplyr::if_else(.data$AVISIT == "BASELINE", "Y", NA_character_),
       ONTRTFL = dplyr::if_else(.data$AVISIT == "POST-BASELINE", "Y", NA_character_),
       ANL01FL = "Y",
       SRCSEQ = NA_integer_
-    ) %>%
-    dplyr::mutate(
-      AVALC = dplyr::case_when(
-        .data$PARAMCD %in% paramcd_yn ~ sample(
-          c("Y", "N"), prob = c(0.1, 0.9), size = dplyr::n(), replace = TRUE),
-        !(.data$PARAMCD %in% paramcd_yn) ~ sample(
-          c("Criteria not met", ">3-5ULN", ">5-10ULN", ">10-20ULN", ">20ULN"), size = dplyr::n(), replace = TRUE)
-      )
-    ) %>%
-    dplyr::mutate(
-      AVAL = dplyr::case_when(
-        .data$AVALC == "Y" ~ 1,
-        .data$AVALC == "N" ~ 0,
-        .data$AVALC == "Criteria not met" ~ 0,
-        .data$AVALC == ">3-5ULN" ~ 1,
-        .data$AVALC == ">5-10ULN" ~ 2,
-        .data$AVALC == ">10-20ULN" ~ 3,
-        .data$AVALC == ">20ULN" ~ 4
-      )
     )
+
+  # remove records for parameters with period 2 and not in paramcd_by_period
+  ADHY <- dplyr::filter(ADHY, .data$PARAMCD %in% paramcd_by_period | .data$APERIODC == "PERIOD 1") # nolint
 
   # add baseline variables
   ADHY <- ADHY %>% # nolint
@@ -175,7 +199,7 @@ radhy <- function(ADSL, # nolint
     dplyr::ungroup() %>%
     dplyr::mutate(ADTM = .data$TRTSDTM + .data$ADY)
 
-  # + operation causes that tzone attribute is lost for POSIXct objects
+  # `+` operation causes that tzone attribute is lost for POSIXct objects
   attributes(ADHY$ADTM) <- attributes(ADSL$TRTSDTM)
 
   # order columns and arrange rows; column order follows ADaM_1.1 specification
@@ -193,6 +217,8 @@ radhy <- function(ADSL, # nolint
       "ADY",
       "AVISIT",
       "AVISITN",
+      "APERIOD",
+      "APERIODC",
       "ONTRTFL",
       "SRCSEQ",
       "ANL01FL"